% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.clustering_part1.R
\name{sample_dbscan}
\alias{sample_dbscan}
\alias{plot.sample_dbscan}
\title{Predict empty partitions and single target clusters in the samples (deep
workflow)}
\usage{
sample_dbscan(
  sample.subquality,
  sample.table,
  referenceDB,
  file.location = "."
)

\method{plot}{sample_dbscan}(x, ..., sample = "all", save.plot = FALSE, format = "png", dpi = 300)
}
\arguments{
\item{sample.subquality}{an object of class \code{read_sample}, inherited
from \code{\link{read_sample}}.}

\item{sample.table}{object of class \code{sample_table}, inherited from
\code{\link{read_sampleTable}}.}

\item{referenceDB}{an object of class \code{reference_dbscan}, inherited
from \code{\link{reference_dbscan}}}

\item{file.location}{character. Full path name to reference and sample
files location. The default corresponds to the working directory,
(\code{\link[base]{getwd}}). Tilde expansion (see
(\code{\link[base]{path.expand}})) is performed.}

\item{x}{an object of class \code{sample_dbscan}}

\item{...}{Arguments to be passed to methods}

\item{sample}{'all' to show all samples, or a numeric vector indicating
the row number of samples in the project table.}

\item{save.plot}{logical. If TRUE the plots are exported to a file.}

\item{format}{a string indicating the file format for the export.
Available formats: 'eps', 'ps', 'tex', 'pdf', 'jpeg', 'tiff', 'png',
'bmp', 'svg', 'wmf'.}

\item{dpi}{numeric. Image resolution.}
}
\value{
An object of class \code{sample_dbscan} containing a sublist for
  each sample. Each sublist has the following components:
  \item{quality}{quality threshold used in \code{\link{read_sample}}.}
  \item{reference}{reference ID.}
  \item{dbscan}{an object of class \code{dbscan_fast}, inherited from
    \code{\link[dbscan]{dbscan}}.}
}
\description{
This function predicts the single target clusters in the samples using the
DBSCAN analysis of a reference as guide.
}
\examples{
\dontrun{
library(dPCP)

#Find path of sample table and location of reference and input files
sampleTable <- system.file("extdata", "Template_sampleTable.csv",
                     package = "dPCP")

fileLoc <- system.file("extdata",package = "dPCP")

#Read sample table file
sample.table <- read_sampleTable(sampleTable, system = "bio-rad",
                                 file.location = fileLoc)

#Read reference files
ref <- read_reference(sample.table, system = "bio-rad",
                      file.location = fileLoc)

#Read samples files
samp <- read_sample(sample.table, system = "bio-rad",
                    file.location = fileLoc)

#Reference DBSCAN clustering
dbref <- reference_dbscan(ref, sample.table, save.template = FALSE)

#Sample DBSCAN prediction
dbsam <- sample_dbscan(samp, sample.table, dbref, file.location = fileLoc)

plot(dbsam, sample = "all", save.plot = FALSE)
}
}
